feat: add delete confirmation and notifications

+++ src/hooks/useUserManagement.ts
@@ -1,11 +1,9 @@
 import { useState } from "react";
+import { toast } from "sonner";
+import Swal from "sweetalert2";
 
 import { useAppDispatch } from "../store/hooks";
-import {
-  createUserThunk,
-  deleteUserThunk,
-  updateUserThunk,
-} from "../features/users/usersSlice";
+import {createUserThunk, deleteUserThunk, updateUserThunk} from "../features/users/usersSlice";
 
 import type { User } from "../types/user";
 import type { UserFormData } from "../components/users/UserForm";
@@ -52,10 +50,40 @@ function useUserManagement() {
 
 
   const handleDelete = async (user: User) => {
+    const isDarkMode = document.documentElement.classList.contains("dark");
+  const result = await Swal.fire({
+    title: "Delete user?",
+    text: `Are you sure you want to delete ${user.name}?`,
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonText: "delete",
+    cancelButtonText: "Cancel",
+    reverseButtons: true,
+
+    background: isDarkMode ? "#111827" : "#ffffff",
+      color: isDarkMode ? "#e5e7eb" : "#1f2937",
+
+      buttonsStyling: false,
+
+    customClass: {
+        confirmButton:
+          "cursor-pointer rounded-xl bg-red-600 px-5 py-2.5 ml-3 text-sm font-semibold text-white hover:bg-red-700",
+
+        cancelButton:
+          "cursor-pointer ml-2 rounded-xl bg-gray-200 px-5 py-2.5 text-sm font-semibold text-gray-700 hover:bg-gray-300 dark:bg-gray-800 dark:text-gray-200 dark:hover:bg-gray-700",
+      },
+  });
+
+  if (!result.isConfirmed) return;
+
   try {
     await dispatch(deleteUserThunk(user.id)).unwrap();
+
+    toast.success("User deleted successfully");
   } catch (error) {
     console.error("Failed to delete user:", error);
+
+    toast.error("Failed to delete user");
   }
 };
